+++ test/WebJobs.Script.Tests/Extensions/FunctionMetadataExtensionsTests.cs
@@ -2,9 +2,12 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System.IO;
+using System.IO.Abstractions;
+using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Azure.WebJobs.Script.Description;
 using Microsoft.Azure.WebJobs.Script.WebHost.Extensions;
+using Moq;
 using Newtonsoft.Json.Linq;
 using Xunit;
 
@@ -123,7 +126,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Extensions
             };
 
             AddSampleBindings(functionMetadata);
-            var result = await functionMetadata.ToFunctionMetadataResponse(options, string.Empty, null);
+            var result = await functionMetadata.ToFunctionMetadataResponse(options, string.Empty, null, excludeTestData: false);
 
             Assert.Null(result.ScriptRootPathHref);
             Assert.Null(result.ConfigHref);
@@ -131,6 +134,66 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Extensions
 
             var binding = result.Config["bindings"] as JArray;
             Assert.Equal("httpTrigger", binding[0]["type"].Value<string>());
+        }
+
+        [Theory]
+        [InlineData(true)]
+        [InlineData(false)]
+        public async Task ToFunctionMetadataResponse_TestData_Exclusion(bool shouldExcludeTestDataFromApiResponse)
+        {
+            var functionName = "TestFunction1";
+            var functionMetadata = new FunctionMetadata
+            {
+                Name = functionName
+            };
+            var testDataFileName = $"{functionName}.dat";
+            var testDataFilePath = Path.Combine(_testRootScriptPath, testDataFileName);
+            var options = new ScriptJobHostOptions
+            {
+                RootScriptPath = _testRootScriptPath,
+                TestDataPath = testDataFilePath
+            };
+
+            IFileSystem fileSystem = FileUtility.Instance;
+
+            try
+            {
+                var testDataContent = @"
+                                    {
+                                        ""method"": ""POST"",
+                                        ""headers"": [],
+                                        ""body"": ""foo""
+                                    }";
+
+                var memoryStream = new MemoryStream(Encoding.UTF8.GetBytes(testDataContent))
+                {
+                    Position = 0
+                };
+                // Mock the file system
+                var mockFileSystem = new Mock<IFileSystem>();
+                var mockFile = new Mock<FileBase>();
+                mockFileSystem.Setup(f => f.Directory.Exists(It.IsAny<string>())).Returns(true);
+                mockFile.Setup(f => f.Exists(It.Is<string>(path => path.EndsWith(Path.Combine(functionName, "function.json"))))).Returns(true);
+                mockFile.Setup(f => f.Exists(It.Is<string>(path => path.EndsWith(testDataFileName)))).Returns(true);
+                mockFile.Setup(f => f.Open(It.Is<string>(path => path.EndsWith(testDataFileName)), It.IsAny<FileMode>(), It.IsAny<FileAccess>(), It.IsAny<FileShare>()))
+                        .Returns(memoryStream);
+                mockFileSystem.Setup(fs => fs.File).Returns(mockFile.Object);
+                FileUtility.Instance = mockFileSystem.Object;
+
+                AddSampleBindings(functionMetadata);
+                var result = await functionMetadata.ToFunctionMetadataResponse(options, string.Empty, null, excludeTestData: shouldExcludeTestDataFromApiResponse);
+
+                Assert.Equal(functionName, result.Name);
+                Assert.Equal(shouldExcludeTestDataFromApiResponse, result.TestData == null);
+                if (!shouldExcludeTestDataFromApiResponse)
+                {
+                    Assert.True(result.TestData.Trim().Length > 0);
+                }
+            }
+            finally
+            {
+                FileUtility.Instance = fileSystem;
+            }
         }
 
         [Fact]