+++ release_notes.md
@@ -3,3 +3,5 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
+
+- Added the option to exclude test data from the `/functions` endpoint API response (#10943)